[IT-1616] Set correct chart height immediately
Before this change, the height of the svg chart was estimated to create
the initial height of the modal dialog that contained the chart. Then, the
modal dialog slowly transitioned to the correct height. Now the height is
of the container is immediately correctly shown. I visually prefer it this
way. And it also helps me with [IT-1616].