[IT-1616] distributionChart: Only redraw necessary elements on resize
Before this change, when the chart container resized then the old svg
was deleted and a completely new svg was created. Now on resizes, the
svg elements continues to exist and only the chart elements that depend
on the width of the chart are being recreated.

This also fixes a bug: Previously, when the container resizes, the check
mark of the Filter drop down menu was reset to "Descending Median". Even
if a different filter is currently active.

This change also helps me with [IT-1616] because if the old svg gets
deleted on resizes then the download dialog would not be working properly.