[IT-1681] Rework BootStrap & UrlMappings location
In f067be7, the two files were moved out of the default
package. I have moved BootStrap.groovy back into it because
the documentation of Grails 3.3.4 says so here
http://docs.grails.org/3.3.4/guide/single.html#_per_environment_bootstrapping
I have further changed the package of the UrlMappings class
from osm to de.iteratec.osm. This is also in accordance with
the Grails 3.3.4 documentation, as you can see here
http://docs.grails.org/3.3.4/guide/single.html#urlmappings