[IT-1681] Decrease selenium version
Since the grails version upgrade, the selenium dependencies
are subject to a gradle version conflict and version 2.53.1 is
selected for some of the selenium dependencies. This causes a
compile error of the integration tests. The version regression
can be shown with the command
./gradlew dependencyInsight --dependency selenium --configuration testCompile
However, I am unable to find out the root cause of it. The only
solution that I can think of at the moment is to explicitly use
the older 2.53.1 version.