[IT-1681] Geb Test: Don't click on non-existent element
Geb has made the following change in its new version:
https://github.com/geb/issues/issues/501
Previously, nothing would happen if you would call click() on a
non-existent element.

I suppose that the element with had the id pointTwo when our geb
test was written. And at some point, the id changed to pointTwo.id
but our geb test never noticed because clicking on a non-existent
element would not throw an error.

Now that we are using the new geb version, I am also updating the
selector to fix the test.