[IT-1681] Handle type change: GrailsDomainClass to HibernatePersistentEntity
In the new Grails version, the type of the variable domainClass in
_list.gsp has changed from an implementing class of
GrailsDomainClass to the class HibernatePersistentEntity. The
Grails documentation briefly mentions this change here
http://docs.grails.org/latest/guide/upgrading.html#_grails_domain_class_api_deprecated
But that is not very helpful.

The _list.gsp file is part of the fields plugin and also mentions
the change here:
http://grails-fields-plugin.github.io/grails-fields/2.2.3/guide/index.html#_version_2_2_0

Here is the documentation for the old and the new type:
http://docs.grails.org/3.1.5/api/grails/core/GrailsDomainClass.html
http://gorm.grails.org/6.1.x/hibernate/api/org/grails/orm/hibernate/cfg/HibernatePersistentEntity.html
As you can see, both types have the method getPersistentProperties().
Except that getPersistentProperties() of GrailsDomainClass returned
objects of type GrailsDomainClassProperty and the method of
HibernatePersistentEntity now returns objects of type PersistentProperty.

Because HibernatePersistentEntity does not implement getPropertyName()
and PersistentProperty does not implement getNaturalName(), I had to
change these function calls to getDecapitalizedName() and respectively
to getName().

I noticed this type change because the geb specs CsTargetGraphGebSpec
and BrowserAliasGebSpec were broken. Unfortunately, the
getPersistentProperties() of HibernatePersistentEntity return the
persistentProperties in a different order than GrailsDomainClass did.
GrailsDomainClass seems to have returned them in the order in which
they are listed in the static constraints field of the corresponding
domain class. And HibernatePersistentEntity returns them apparently
in an arbitrary order, which is less visually pleasing. But because
the _list.gsp does not affect many of our pages and I know of no
quick fix, I kept it this way and only adjusted the geb tests.