[IT-1681] Correctly assign UserTimings to EventResult as 1:n
The domain class EventResult has declared a hasMany relationship to
UserTimings. hasMany is supposed to model a 1:n relationship, as documented
here: http://docs.grails.org/3.3.2/ref/Domain%20Classes/hasMany.html
The relationship between those two entities is modeled via a join table
in the schema.

In the createData method of the geb spec, the same userTimingsForJobResult2
list was used in multiple constructor calls of EventResults which
resulted in a many-to-many relationship persisted in the database.
I don't understand why Grails does not correctly enforce the 1:n
relationship. In my understanding, persisting the second EventResult
with the same userTimings should already not have passed validation.

I am now correctly creating the test data. Otherwise, the cleanUpData()
function throws an exception in the new Grails version as described in
c5fe3cd.