[IT-1681] Call geb method on single-element navigator
As described in the commit comment of bd62949, we can
no longer call these geb methods on multi-element
navigators. The navigators graphNameSelect and colorPicker
are selected by the id #graphName respectively #color.
Unfortunately, each time there are two elements with the
same id because part of the page was cloned in
grails-app/assets/javascripts/charts/chart-adjustment.js.
Disregarding if it is a good idea two have the same id
twice, I adjusted the geb specs correspondingly.