[IT-2140] Replace unicode compare arrow <-->
We added the unicode compare arrow in f2547da. As it turns out,
the shape of the arrow depends on the font used and that font
can differ between browsers, operating systems, etc. The old shape
looked ugly with my Chrome and with Oles Firefox. It only looked good
with Oles Chrome. The unicode arrow that we now added looked good
with both fonts that we tried.

Co-authored-by: ole.wellnitz@iteratec.de <ole.wellnitz@iteratec.de>