[IT-2203] Use correct spin.js dependency
During the migration from bower to npm, the dependency to version ^2.0.1
of spin.js was dropped in exchange with version ^0.0.1 of spin.
This happened in commits 281087e, 171c5c1 and 29c5ecc. This lead to
the spin icon not being centered in its respective div. Instead, it was
positioned in the upper left corner.