Fix bidirectional oneToMany field
Just recently in 9fed303, I had changed
persistentProperty.getReferencedDomainClass().getName() to
persistentProperty.getAssociatedEntity().getName(). Now, I am changing
getName() to getDecapitalizedName(). The getName() method returns
the name including the whole package prefix. getDecapitalizedName()
omits the package name. The change to getDecapitalizedName() has become necessary with
the upgrade of the Grails version from 3.3.2 to 3.3.5. That version
upgrade broke BrowserGebSpec. Because BrowserGebSpec used
oneToMany/_bidirectional.gsp. In _bidirectional.gsp, the variable
propertyClass was used, which is being set here. In _bidirectional.gsp
it is used as an argument to the createLink groovy Tag. It looks like,
in Grails 3.3.2, the tag could handle the package prefix and still
create a correct link. In Grails 3.3.5 however, the package prefix of
propertyClass lead to a broken link.