Our passwords are always encoded
The line that I changed was originally generated by the command
`grails s2-quickstart`, as you can see here:
https://grails-plugins.github.io/grails-spring-security-core/3.2.x/index.html#3-create-the-user-and-role-domain-classes
That generated code had to anticipate the possibility that the
passwordEncoder bean is not defined, which is the case if digest
authentication is used and the configuration attribute
digest.useCleartextPasswords is set to true. As described here:
https://grails-plugins.github.io/grails-spring-security-core/3.2.x/index.html#encodepassword
Neither are we using digest authentication nor do we ever want to use
cleartext passwords.

That's why, in our case, the passwordEncoder bean is always going to be
defined and the ? operator is unnecessary. I am changing the code to
unambiguously reflect that fact.