[IT-2264] Improve bar label transition between average & median
During the transition from a chart with selected aggregation value
average to median or in the other direction, the label of a bar
could have shown "not definedms". To avoid this, I am now returning
an empty string if d.value is not defined.

Additionally, we no longer set the bar label in a transition. The change
of the label happened instantaneously at the start of the transition
anyways. The only problem was that getComputedTextLength() which was
called inside the same transition would return 0 for the text that
was changed in the transition. This resulted in the opacity value
sometimes not being set to 0 when it should have been.