[IT-2464] docker: Allow user independent usage
Systems like OpenShift do not gurantee that a image is actually run with the user specified in it.
This caused problems as the image relied on using the $HOME variable and being user 'osm' with specific rights.

Now the user only needs to be in the user group and the rest works with arbitrary users.